+++ project/src/components/login-form/login-form.module.css
@@ -2,6 +2,11 @@
   position: relative;
 }
 
+.invalidInput {
+  border: 1px solid red;
+  outline: none;
+}
+
 .errorMessage {
   position: absolute;
   left: 2px;
